requirements.txt
pyaudio>=0.2.14
vosk>=0.3.45
httpx>=0.25.0
PyQt6>=6.6.0
pyopenxr>=1.0.0
numpy>=1.24.0
sounddevice>=0.4.6
Pillow>=10.0.0

